vinyl: allocate records for metadata log dynamically
Currently, we use a fixed size buffer, which can accommodate up to 64
records. With the single memory level it can easily overflow, as we
create a slice for each range on dump in a single transaction, i.e. if
there are > 64 ranges in an index, we may get a panic. So this patch
makes vylog use a list of dynamically allocated records instead of a
static array.