vinyl: move run min/max lsn from vylog to index file
This reverts commits
  fbe16d3d9555 ("vinyl: store min_lsn and max_lsn in vylog")
  19a5b3dd74a8 ("vinyl: rework run min/max lsn calculation")

The former single memory level design required knowledge of max LSN of
each run. Since this information can't be extracted from the run file in
general (the newest key might have been deleted by compaction), we added
it to the metadata log. Thanks to the concept of slices, this is not
needed any more, so we can calculate min/max LSN as min/max over all
statements that have been written to the run file (as we used to) and
move them back to the index file.