sql: refactor box.sql.execute
lua_sql_execute executes query and pushes
results on lua stack.
But we want for ability to send sql results via
iproto. To reach the goal it is need to return
sql results in some intermediate format, that can
be converted in lua tables or iproto messages.

This patch implements box_sql_execute, that
returns 'struct port' with result tuples list.
struct port can be dumped in obuf and sent via iproto,
or printed to Lua.