Add support for space:format() to net.box
Getting the space format should be safe, as it is tied to schema_id,
and net.box makes sure that schema_id stays consistent.

It means that when you receive a tuple from net.box, you may be sure
that its space format is consistent with the remote.

In context of #2402